More function movement to utils/init.lua and util/links.lua
